+++ src/WallThickness/CMakeLists.txt
@@ -5,9 +5,11 @@
 set(header_path "${BSMPT_SOURCE_DIR}/include/BSMPT/WallThickness")
 set(header
 	${header_path}/WallThicknessLib.h
+	${header_path}/WallThicknessCommon.h
 )
 set(src
 	WallThicknessLib.cpp
+	WallThickness1D.cpp
 )
 
 add_library(LibWallThickness ${header} ${src})
